Change TEE_GetNextPersistentObject() objectIDLen type to uint32_t
The type of the objectIDLen was declared as size_t. However, the
TEE Internal Core API Specification v1.1 defines the type of the
TEE_GetNextPersistentObject() objectIDLen parameter as uint32_t.

+++ internal_api/tee_storage_api.h
@@ -314,7 +314,7 @@ TEE_Result TEE_StartPersistentObjectEnumerator(TEE_ObjectEnumHandle objectEnumer
  */
 TEE_Result TEE_GetNextPersistentObject(TEE_ObjectEnumHandle objectEnumerator,
 				       TEE_ObjectInfo *objectInfo, void *objectID,
-				       size_t *objectIDLen);
+				       uint32_t *objectIDLen);
 
 /*
  * ## Data Stream Access Functions ##
